fix: cerrar los hallazgos de la auditoría integral del backend
Los tres de mayor severidad primero.

`.gitignore` sólo cubría el archivo `.env` exacto, así que una copia hecha antes
de un cambio grande —con los mismos secretos dentro— quedaba esperando a que un
`git add -A` la publicara. El patrón pasa a ser `.env.*`, con `.env.example`
exceptuada por ser la plantilla.

`src/app.module.ts` importaba un directorio entero que no estaba versionado, y
`prisma.service.ts` un archivo suelto igual: un clon limpio de la rama no
compilaba. Se versionan el módulo de persistencia, la telemetría y tres
migraciones, entre otros.

Y seis sentencias crudas corrían fuera de transacción sobre tablas con RLS
forzada. `set_config(..., true)` parece local, pero una vez que una conexión la
ha usado el parámetro queda DEFINIDO con cadena vacía para el resto de su vida:
deja de ser NULL, la política intenta `''::bigint` y aborta con 22P02. Con
`WORKER_ROLE=ALL` los trabajos de fondo comparten pool con las peticiones, así
que les tocaban conexiones ya contaminadas y el orquestador lo enterraba bajo un
backoff exponencial. Reproducido contra Postgres real en una prueba nueva.

Además, el motor publicaba `score: null` con `status: SUCCEEDED` cuando una
acción traía un valor no numérico: `Number()` daba NaN, NaN atravesaba las sumas
sin quejarse y `JSON.stringify` lo convertía en null. El evaluador de
expresiones ya cerraba esa puerta; faltaba la de la configuración del nodo y el
payload de las acciones.

La idempotencia cerraba por `id` a secas, de modo que un titular cuya ejecución
se pasó del lease pisaba la reserva que otra petición ya había reclamado.
`reserve()` devuelve ahora el lease como comprobante de propiedad, y el env
schema exige que ese lease supere la decisión más larga admisible, con lo que la
carrera deja de ser alcanzable por configuración.

Se añade ESLint con información de tipos —lo único que ve una promesa sin
esperar— más un suelo de cobertura, y pruebas para los cuatro módulos que no
tenían ninguna: health, live-execution, security-review y tutorials.

Gates: lint 0 errores · typecheck · build · 1096 pruebas · 106 de integración
contra Postgres real · contrato OpenAPI regenerado · docs.

+++ src/common/security/authentication.guard.ts
@@ -160,7 +160,7 @@ export class AuthenticationGuard implements CanActivate {
     const requested = this.header(request, 'x-tenant-id');
     if (!requested) {
       // Selecting a tenant implicitly is only unambiguous for single-tenant clients.
-      if (allowed.length === 1) return allowed[0] as bigint;
+      if (allowed.length === 1) return allowed[0];
       throw new DomainException(
         'INVALID_SECURITY_CONTEXT',
         'x-tenant-id is required for clients authorised on multiple tenants',
